% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_effects_plots.R
\name{compute_grand_mean}
\alias{compute_grand_mean}
\title{Computes the grand mean of a data frame}
\usage{
compute_grand_mean(df, effects_var_vec, response_var, weighted = TRUE)
}
\arguments{
\item{df}{the data frame.}

\item{effects_var_vec}{the vector of factors to group by for the average.}

\item{response_var}{the response variable.}

\item{weighted}{book a boolean to determine if the grand mean average is weighted by effect
or computed unweighted.
If [], this function will just return the average of all data points. This is one of
the only functions that can take an empty [] effects_var_vec.

Missing values are removed from the averaging, making it necessary to average all of the groups}
}
\description{
The grand mean is an average of each of the groups averages, so we
compute the grand mean according to the grouping specified by effects_var_vec
}
