% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_plots.R
\name{compute_mean_df}
\alias{compute_mean_df}
\title{Basic function to group by all factors and compute a mean}
\usage{
compute_mean_df(df, effects_var_vec, response_var)
}
\description{
This is called by other methods
}
