% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_plot.R
\name{effects_compute_df}
\alias{effects_compute_df}
\title{Produce a table of |effects| sizes for factors up to the specified interaction depth}
\usage{
effects_compute_df(
  df,
  effects_var_vec,
  response_var,
  interaction_depth = 2,
  max_value_only = TRUE
)
}
\description{
Produce a table of |effects| sizes for factors up to the specified interaction depth
}
