% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_plot.R
\name{effects_one_term_df}
\alias{effects_one_term_df}
\title{Compute effect differences of main effects in effects df table format}
\usage{
effects_one_term_df(
  df,
  effects_var_vec,
  response_var,
  max_value_only = TRUE,
  plot_NA_values = FALSE,
  alpha = 0.05
)
}
\description{
Table format: ("Factors", "Effect", "abs(Effect)", "Max_Level", "Min_Level")
}
