% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_plot.R
\name{effects_two_term_df}
\alias{effects_two_term_df}
\title{Compute effect differences of two_term interactionseffects in effects df table format}
\usage{
effects_two_term_df(
  df,
  effects_var_vec,
  response_var,
  max_value_only = TRUE,
  plot_NA_values = FALSE,
  alpha = 0.05
)
}
\description{
Table format: ("Factors", "Effect", "abs(Effect)", "Max_Level", "Min_Level")
This assumes we have at least two factors
}
