% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_plot.R
\name{effects_two_term_level_df}
\alias{effects_two_term_level_df}
\title{Compute all of the effect sizes for all of the different pairings of one combination
of Factor variables}
\usage{
effects_two_term_level_df(
  df,
  effect_var_1,
  effect_var_2,
  response_var,
  max_value_only = FALSE
)
}
\arguments{
\item{df}{the data frame to compute effects on}

\item{effect_var_1}{the list of the two-term effect interactions}

\item{response_var}{the response variable}

\item{max_value_only}{if TRUE, only give the maximum value for each interaction
If FALSE, compute all values
Use main_effects_hsd_compute_df() to compute the effect sizes for one-term level
changes}
}
\description{
Compute all of the effect sizes for all of the different pairings of one combination
of Factor variables
}
