% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{eh_position_legend}
\alias{eh_position_legend}
\title{Customizes the position of the legend
This replaces the "outside" argument in plots that customizes legend positioning}
\usage{
eh_position_legend(plot_obj, position)
}
\arguments{
\item{plot_obj}{the current plot object}

\item{position}{The position of the legend, conforming with ggplot2
specifications for position.
It can be a word like "bottom" or c(xpos,ypos) vector from the bottom-right corner of the plot}
}
\value{
plot_obj a plot object with the repositioned legend
}
\description{
Customizes the position of the legend
This replaces the "outside" argument in plots that customizes legend positioning
}
