% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{eh_title}
\alias{eh_title}
\title{Replaces title of a plot.
Title can contain new lines with escape characters for multi-line plots.}
\usage{
eh_title(plot_obj, titlestr)
}
\arguments{
\item{plot_obj}{the current plot object}

\item{titlestr}{the desired title text}
}
\value{
plot_obj a plot object with the modified title
}
\description{
Replaces title of a plot.
Title can contain new lines with escape characters for multi-line plots.
}
