% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{eh_window_limits}
\alias{eh_window_limits}
\title{Zoom the plot to the desired x and y axis limits}
\usage{
eh_window_limits(plot_obj, xlim_vec, ylim_vec)
}
\arguments{
\item{plot_obj}{the current plot object}

\item{xlim}{the x-axis limits as (xmin, xmax)}

\item{ylim}{the y-axis limits as (ymin, ymax)}
}
\value{
a plot_object with zoomed-in limits.
}
\description{
It uses coord_cartesian() from ggplot2 so that data outside the window is not
thrown out (which actually influences how
boxplots can look)
}
