% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{eh_x_label}
\alias{eh_x_label}
\title{Replaces x-axis label of a plot}
\usage{
eh_x_label(plot_obj, xlabel)
}
\arguments{
\item{plot_obj}{the current plot object}

\item{xlabel}{the desired x axis label text}
}
\value{
plot_obj a plot object with the modified x-axis label
}
\description{
Replaces x-axis label of a plot
}
