% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{eh_y_label}
\alias{eh_y_label}
\title{Replaces y-axis label of a plot}
\usage{
eh_y_label(plot_obj, ylabel)
}
\arguments{
\item{plot_obj}{the current plot object}

\item{ylabel}{the desired y axis label text}
}
\value{
plot_obj a plot object with the modified y-axis label
}
\description{
Replaces y-axis label of a plot
}
