% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_custom.R
\name{geom_text_custom}
\alias{geom_text_custom}
\title{Custom geom text custom rendering geom}
\usage{
geom_text_custom(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  geom = geom_text_custom_obj
)
}
\description{
Taken from https://stackoverflow.com/questions/49965758/change-geom-texts-default-a-
(continued) legend-to-label-string-itself
}
