% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_custom.R
\name{get_custom_draw_key}
\alias{get_custom_draw_key}
\title{Custom Geom Text Object draw key method}
\usage{
get_custom_draw_key(var_key_lvl, cols_vec)
}
\description{
Taken from https://stackoverflow.com/questions/49965758/change-geom-texts-default-a-
(continued) legend-to-label-string-itself
var_key = block_point_levels, (or some levels)
cols = scales::hue_pal()(length(var_key))
Custom Geom Text Object draw key method
}
\details{
Taken from https://stackoverflow.com/questions/49965758/change-geom-texts-default-a-
(continued) legend-to-label-string-itself
var_key = block_point_levels, (or some levels)
cols = scales::hue_pal()(length(var_key))
}
