% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_custom.R
\name{get_custom_text_object}
\alias{get_custom_text_object}
\title{Get Custom geom_text object to have legends contain symbols in graph}
\usage{
get_custom_text_object(var_key, cols = (scales::hue_pal())(length(var_key)))
}
\description{
Adapted from https://stackoverflow.com/questions/49965758/change-geom-texts-default-a-
(continued) legend-to-label-string-itself
Produces a custom geom_text object that is used to enhance plots. Placed here to avoid
the replication in individual plots.
}
