% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Returns the current plot data frame as a data frame object}
\usage{
get_plot_data(plot_obj)
}
\arguments{
\item{plot_obj}{the current plot object}
}
\value{
plot_df the data frame used to produce the plot.
}
\description{
Return of plot data used for additional plotting. Data frames may have extra rows and columns
needed for plotting and plot aesthetics.
}
