% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{get_title}
\alias{get_title}
\title{Returns the current plot title as a string}
\usage{
get_title(plot_obj)
}
\arguments{
\item{plot_obj}{the current plot object}
}
\value{
title_str the title of the plot as a string.
}
\description{
Title return used for testing plots.
}
