% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{get_x_label}
\alias{get_x_label}
\title{Returns the current plot x-axis label as a string}
\usage{
get_x_label(plot_obj)
}
\arguments{
\item{plot_obj}{the current plot object}
}
\value{
x_str the x-axis label of the plot as a string.
}
\description{
X-Axis label return used for enhancing x-axes labels
}
