% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{get_y_label}
\alias{get_y_label}
\title{Returns the current plot y-axis label as a string}
\usage{
get_y_label(plot_obj)
}
\arguments{
\item{plot_obj}{the current plot object}
}
\value{
y_str the y-axis label of the plot as a string.
}
\description{
Y-Axis label return used for enhancing y-axes labels
}
