% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_plots.R
\name{interact_compare_produce_plot}
\alias{interact_compare_produce_plot}
\title{Produces a main effects plot}
\usage{
interact_compare_produce_plot(
  effects_all_df,
  labels_vec,
  effects_var_vec,
  treatment_var,
  response_var,
  plot_NA_values = FALSE,
  comp_func = mean,
  ...
)
}
\arguments{
\item{effects_all_df}{the data frame with the data and all of the effects}

\item{labels_vec}{the vector of the labels}

\item{effects_var_vec}{the list of colum names of the effect factors, specified in order}

\item{treatment_var}{the treatment col to compare in all of these plots}

\item{response_var}{the column name of the response variable}

\item{plot_NA_values}{if TRUE, plot missing values. FALSE by default.}

\item{comp_func}{mean by default. The computation function used to aggregate the multiple
values into a single value.}
}
\value{
plot_obj a plot object
}
\description{
Produces a main effects plot
}
