% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_effects_plots.R, R/youden_plots.R
\name{main_effects_mar_produce_plot}
\alias{main_effects_mar_produce_plot}
\title{Produces a main effects plot with a pre-computed intermediary data frame}
\usage{
main_effects_mar_produce_plot(
  effects_all_df,
  labels_vec,
  variable_fac,
  value_fac,
  response_var
)

main_effects_mar_produce_plot(
  effects_all_df,
  labels_vec,
  variable_fac,
  value_fac,
  response_var
)
}
\arguments{
\item{effects_all_df}{the data frame with the data, already aggregated}

\item{variable_fac}{the column name representing the column variables}

\item{value_fac}{the column name representing the values of the factors in
column `variable_fac`}

\item{response_var}{the column name of the response variable}

\item{label_vec}{the column names of the labels.}

\item{df}{the data frame with the data, already aggregated}
}
\value{
plot_obj a plot object

plot_obj a plot object
}
\description{
Produces a main effects plot with a pre-computed intermediary data frame

Produces a main effects plot with a pre-computed intermediary data frame
}
