% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_effects_plots.R
\name{main_effects_produce_plot}
\alias{main_effects_produce_plot}
\title{Produces a main effects plot}
\usage{
main_effects_produce_plot(
  effects_all_df,
  labels_vec,
  mean_val,
  effects_var_vec,
  response_var,
  sd_df = NULL,
  plot_NA_values = FALSE,
  bar_ci = FALSE,
  alpha = 0.05,
  bar_sd = FALSE
)
}
\arguments{
\item{effects_all_df}{the data frame with the data}

\item{labels_vec}{the vector of all of the lables}

\item{mean_val}{the grand mean.}

\item{effects_var_vec}{the list of column names of the effect factors, specified in order}

\item{response_var}{the column name of the response variable}

\item{plot_NA_values}{FALSE by default. If true, plot the main effects that are missing.}

\item{bar_ci}{If TRUE, plot bars that are the standard error.}

\item{alpha}{the confidence level to produce a 1 - alpha confidence interval}

\item{bar_sd}{FALSE by default. If TRUE, plot the standard deviation bars.
FALSE by default. Here sdn is computed similar to a standard error on the mean.}
}
\value{
plot_obj a plot object
}
\description{
Produces a main effects plot
}
