% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_effects_plot}
\alias{mizr_effects_plot}
\title{Produce effects Plot}
\usage{
mizr_effects_plot(
  df,
  effects_var_vec,
  response_var,
  interaction_depth = 2,
  max_value_only = TRUE
)
}
\arguments{
\item{df}{the data frame}

\item{effects_var_vec}{the vector of effects or factors to compute the effects of}

\item{response_var}{the response column to plot on the y axis;}

\item{interaction_depth}{2. The number of interaction terms to compute.
Currently 2 is only supported value.}

\item{max_value_only}{if TRUE, only give the maximum value for each effect or interaction
(For multiple level effects); if FALSE, give all effects}
}
\value{
mizr_plot a plot object with this plot to display
}
\description{
Produce effects Plot
}
