% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_histogram_plot}
\alias{mizr_histogram_plot}
\title{Plots a histogram}
\usage{
mizr_histogram_plot(
  df,
  x_val,
  bins = NULL,
  binwidth = NULL,
  closed = "left",
  add_text = TRUE
)
}
\arguments{
\item{df}{the data frame}

\item{x_val}{the column name for the x row}

\item{bins}{the number of bins to use Please specify one of bins or binwidth.}

\item{binwidth}{the width of each bin. Please specify one of bins or binwidth.}

\item{closed}{"left" closed by default. Can have "right" closed histograms.}

\item{add_text}{if TRUE, add the number count of each bin as a text label}
}
\value{
mizr_plot a plot object
}
\description{
Plots a histogram
}
