% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_interact_compare_plot}
\alias{mizr_interact_compare_plot}
\title{Produces a two-term interaction plot with respect to the treatment_var comparison factor}
\usage{
mizr_interact_compare_plot(
  df,
  effects_var_vec,
  treatment_var,
  response_var,
  bar_ci = FALSE,
  plot_NA_values = FALSE,
  comp_func = mean,
  ...
)
}
\arguments{
\item{df}{the data frame with the data}

\item{effects_var_vec}{the list of colum names of the effect factors, specified in order}

\item{treatment_var}{the treatment col to compare in all of these plots}

\item{response_var}{the column name of the response variable}

\item{bar_ci}{if TRUE, include confidence intervals; hide if FALSE.}

\item{plot_NA_values}{if TRUE, plot missing values (and keep columns and rows
with missing values). Hide if FALSE.}

\item{comp_func}{the computation function used to aggregate multiple values
into a single value. Defaults to the arithmetic mean.}
}
\value{
mizr_plot a plot object
}
\description{
Produces a two-term interaction plot with respect to the treatment_var comparison factor
}
