% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_lag_plot}
\alias{mizr_lag_plot}
\title{The Lag Plot of Filliben's Fourplots}
\usage{
mizr_lag_plot(
  df,
  response_var,
  run_var = NULL,
  group_fac_vec = NULL,
  label_facets_with_variables = TRUE
)
}
\arguments{
\item{df}{the data frame}

\item{response_var}{the variable to plot}

\item{run_var}{the variable with the run number, or NULL to use the row index number}

\item{group_fac_vec}{the factor vector to group variables. NULL by default.}

\item{label_facets_with_variables}{if TRUE, gives the variable (as well as value) label
to each facet. TRUE by default.}
}
\value{
mizr_plot mizr_plot the plot object
}
\description{
Produces a lag plot comparing the response to the previous response's value
}
