% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_frames.R
\name{mizr_long_to_wide}
\alias{mizr_long_to_wide}
\title{Converts a long data frame to a wide data frame.}
\usage{
mizr_long_to_wide(input_df, factor_col_fac, response_col, trial_id = NULL)
}
\arguments{
\item{input_df}{the long data frame to make wide}

\item{factor_col_fac}{the column that are the factors}

\item{response_col}{the name of the response column}

\item{trial_id}{trial_id column. NULL if no trial_id column exists.}
}
\value{
output_df the wide data frame
}
\description{
Uses the reshape2 package
}
