% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_main_effects_hsd_plot}
\alias{mizr_main_effects_hsd_plot}
\title{Produces Honest Significant Differences (HSD) Confidence Intervals on the Effects}
\usage{
mizr_main_effects_hsd_plot(
  df,
  effects_var_vec,
  response_var,
  plot_NA_values = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{df}{the data frame with the data}

\item{effects_var_vec}{the list of colum names of the effect factors, specified in order}

\item{response_var}{the column name of the response variable}

\item{plot_NA_values}{if TRUE, plot missing values. FALSE by default.}

\item{alpha}{The alpha used for statistical significance tests and to provide (1 - alpha)
confidence intervals}
}
\value{
mizr_plot a plot object
}
\description{
Uses Tukey's Honest Significant Differences (HSD) to compute the confidence intervals to
prevent Type I inflaction error when comparing the error bars
}
