% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_main_effects_mar_plot}
\alias{mizr_main_effects_mar_plot}
\title{Produces a marginal main effects plot}
\usage{
mizr_main_effects_mar_plot(df, variable_fac, value_fac, response_var)
}
\arguments{
\item{df}{the data frame with the data}

\item{variable_fac}{the column name representing the column variables}

\item{value_fac}{the column name representing the values of the factors in
column `variable_fac`}

\item{response_var}{the column name of the response variable}
}
\value{
mizr_plot a plot object
}
\description{
Produces a marginal main effects plot
}
