% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_main_effects_plot}
\alias{mizr_main_effects_plot}
\title{Produces a main effects plot}
\usage{
mizr_main_effects_plot(
  df,
  effects_var_vec,
  response_var,
  plot_NA_values = FALSE,
  bar_ci = FALSE,
  alpha = 0.05,
  bar_sd = FALSE,
  weighted = FALSE
)
}
\arguments{
\item{df}{the data frame with the data}

\item{effects_var_vec}{the list of colum names of the effect factors, specified in order}

\item{response_var}{the column name of the response variable}

\item{plot_NA_values}{if TRUE, plot missing values. FALSE by default}

\item{bar_ci}{If TRUE, plot bars that are the standard error.}

\item{alpha}{The alpha used for statistical significance tests and to provide (1 - alpha)
confidence intervals}

\item{bar_sd}{If TRUE, plot bars that are the standard deviation.}

\item{weighted}{TRUE if the main effect plots should be weighted by effect grouping and then
averaged or FALSE if each category should be averaged separately}
}
\value{
mizr_plot a plot object
}
\description{
Produces a main effects plot
}
