% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{mizr_make_formula}
\alias{mizr_make_formula}
\title{Construct a formula string of response_var ~ effects_var_vec, interaction depth specified}
\usage{
mizr_make_formula(effects_var_vec, response_var, interaction_depth = 1)
}
\arguments{
\item{effects_var_vec}{the vector of effect variables to make the formula for}

\item{response_var}{the response variable for the formula}

\item{interaction_depth}{the desired depth of interactions to compute in the formula. The
default is 1, which only computes the effects}
}
\value{
formula_str, a string of the desired formula
}
\description{
Constructs a forumla string of response_var ~ effects_var_vec,
which includes all of the variables
and all of the couplings (interactions) of all combinations all the way down to the
interaction depth specified
}
