% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_normal_prob_plot}
\alias{mizr_normal_prob_plot}
\title{Produce a normal probability plot}
\usage{
mizr_normal_prob_plot(df, y_var)
}
\arguments{
\item{df}{the data frame}

\item{y_var}{the name of the variable to produce the probability plot for}
}
\value{
mizr_plot a plot object
}
\description{
Produce a normal probability plot
}
