% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_order_plot}
\alias{mizr_order_plot}
\title{Produce an Order Plot of the 10-step series}
\usage{
mizr_order_plot(
  df,
  fac_vec,
  response_var,
  digits = 2,
  remove_NA_factors = FALSE
)
}
\arguments{
\item{df}{the data frame}

\item{fac_vec}{the factors to plot on the x column}

\item{response_var}{the response column to plot on the y axis.}

\item{digits}{the number of digits to round values to}

\item{remove_NA_factors}{if TRUE, remove all values where one or more of the factors
is NA. FALSE by default}
}
\value{
mizr_plot a plot object
}
\description{
Produce an Order Plot of the 10-step series
}
