% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_run_plot}
\alias{mizr_run_plot}
\title{The Run Plot of Filliben's Fourplots}
\usage{
mizr_run_plot(
  df,
  response_var,
  group_fac_vec = NULL,
  run_var = NULL,
  label_facets_with_variables = TRUE
)
}
\arguments{
\item{df}{the data frame}

\item{response_var}{the variable to plot}

\item{run_var}{the variable with the run number, or NULL to use the row index number}

\item{label_facets_with_variables}{if TRUE, gives the variable (as well as value) label
to each facet. TRUE by default.}

\item{group_var_vec}{the factor variables to group by for a run plot faceted by group}
}
\value{
mizr_plot mizr_plot the plot object
}
\description{
Produces a run plot comparing the run number to the response variable
}
