% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_scatter_doe_plot}
\alias{mizr_scatter_doe_plot}
\title{Produces a scatter plot that colors and shapes by the treatment factor}
\usage{
mizr_scatter_doe_plot(
  df,
  fac_vec,
  response_var,
  treatment_fac = NULL,
  doe_plot_style = "enhanced",
  plot_NA_values = FALSE
)
}
\arguments{
\item{df}{the data frame}

\item{response_var}{the response variable to plot on the y axis}

\item{treatment_fac}{the treatment factor}

\item{doe_plot_style}{Specifies the Plot Style. Use "plain" to mimic the DOE Scatter Plot
in dataplot, and "enhanced" to use an enhanced version.
"boxplot" to place the boxplot in front of the points, and "sd" to replace the boxplot
To produce a mean and standard deviation rather than quantiles. Defaults to "enhanced".}

\item{plot_NA_values}{FALSE by default. If TRUE, plot all values with a factor of a
NA value.}

\item{x_var}{the factor to plot on the x column}
}
\value{
mizr_plot the plot object
}
\description{
Produces a scatter plot that colors and shapes by the treatment factor
}
