% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_scatter_effect_plot}
\alias{mizr_scatter_effect_plot}
\title{Produces a scatter plot that does a main effect of x_var by response_var}
\usage{
mizr_scatter_effect_plot(df, x_var, response_var, comp_func = mean, ...)
}
\arguments{
\item{df}{the data frame}

\item{x_var}{the factor to plot on the x column}

\item{response_var}{the response variable to plot on the y axis}
}
\value{
mizr_plot the plot object
}
\description{
Produces a scatter plot that does a main effect of x_var by response_var
}
