% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_scatter_treatment_plot}
\alias{mizr_scatter_treatment_plot}
\title{Produces a scatter plot that colors and shapes by the treatment factor}
\usage{
mizr_scatter_treatment_plot(
  df,
  x_var,
  treatment_fac,
  response_var,
  shape_style = "point"
)
}
\arguments{
\item{df}{the data frame}

\item{x_var}{the factor to plot on the x column}

\item{treatment_fac}{the treatment factor}

\item{response_var}{the response variable to plot on the y axis}

\item{shape_style}{a string indicating if the shape should be points ('point') or textual
symbols 'txt'. Defaults to point.}
}
\value{
mizr_plot the plot object
}
\description{
Produces a scatter plot that colors and shapes by the treatment factor
}
