% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_plots.R
\name{mizr_set_theme}
\alias{mizr_set_theme}
\title{Set ggplot theme to default mizr theme
Sets the theme to the default mizr theme
The theme can be further updated with:
old_theme <- theme_set(theme_get() + <updates>)}
\usage{
mizr_set_theme(base_size = 13)
}
\value{
old_theme the old_theme before the theme change, as is consistent
with ggplot themees
}
\description{
Set ggplot theme to default mizr theme
Sets the theme to the default mizr theme
The theme can be further updated with:
old_theme <- theme_set(theme_get() + <updates>)
}
