% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_sign_test_plot}
\alias{mizr_sign_test_plot}
\title{Produces a block rank plot}
\usage{
mizr_sign_test_plot(
  df,
  x_var_vec,
  treatment_var,
  response_var,
  higher_is_better = TRUE,
  use_margins = FALSE,
  include_equal_comparisons = TRUE,
  alpha = 0.05,
  p_adj = "bonferonni",
  low_color = "#FF0000",
  mid_color = "#FFFFFF",
  high_color = "#0074FF",
  tile_text_size = rel(2),
  comp_func = mean,
  ...
)
}
\arguments{
\item{df}{the data frame}

\item{x_var_vec}{the vector of columns to display on the x-axis.
Multiple factors can be combined to enumerate out the different values.}

\item{treatment_var}{the blocking factor or effect to plot separately as different symbols}

\item{response_var}{the response column to plot on the y axis;}

\item{higher_is_better}{TRUE if higher scores are better, FALSE if lower scores are
better}

\item{use_margins}{if TRUE, add marginal sum columns}

\item{alpha}{type I significance}

\item{p_adj}{either "none" or "bonferonni". "none" does not adjust the p-values for Type I
inflation while "bonferonni" uses a bonferonni correction to the Type I error}

\item{low_color}{the color for the lowest value.}

\item{mid_color}{the color for the midpoint value.}

\item{high_color}{the color for the highest value.}

\item{tile_text_size}{the size of the text within the plot. Often specified
as rel(Y) or a relative text size}

\item{comp_func}{the computation function used to aggregate multiple values
into a single value. Defaults to the arithmetic mean.}
}
\value{
mizr_plot a plot object with this plot to display
}
\description{
Produces a block rank plot
}
