% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_tile_table_count_plot}
\alias{mizr_tile_table_count_plot}
\title{Produce a Colored Table of all counts of the number of attributes}
\usage{
mizr_tile_table_count_plot(
  df,
  x_fac_vec,
  y_fac_vec,
  response_name = "count",
  low_color = "#FFFFFF",
  mid_color = "#EEEEEE",
  high_color = "#0074FF",
  midpoint = 1,
  digits = 2,
  use_margins = FALSE,
  tile_text_size = rel(1.6),
  remove_NA_factors = FALSE
)
}
\arguments{
\item{df}{the data frame}

\item{x_fac_vec}{the vector of factors to go on the x-axis.}

\item{y_fac_vec}{the vector of factors to go on the y-axis.}

\item{low_color}{the color for the lowest value.}

\item{mid_color}{the color for the midpoint value.}

\item{high_color}{the color for the highest value.}

\item{midpoint}{the numerical value of the middle point to color mid_color}

\item{digits}{the number of digits to round numbers in the plot to. Defaults to 2.}

\item{use_margins}{if TRUE, provide row and column margin aggregates. FALSE by
default.}

\item{tile_text_size}{the size of the text in tiles.
can be a number or a rel(1.0) for relative sizes.}

\item{remove_NA_factors}{if TRUE, remove all attributes with a NA factor. FALSE by
default.}

\item{response_var}{the name to display for the count response. Defaults to "count"}
}
\value{
mizr_plot the plot object
}
\description{
Produce a Colored Table of all counts of the number of attributes
}
