% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_tile_table_plot}
\alias{mizr_tile_table_plot}
\title{Produce a Colored Table of all factors}
\usage{
mizr_tile_table_plot(
  df,
  x_fac_vec,
  y_fac_vec,
  response_var,
  low_color = "#FF0000",
  mid_color = "#FFFFFF",
  high_color = "#0074FF",
  midpoint = 0,
  digits = 2,
  use_margins = FALSE,
  tile_text_size = rel(1.6)
)
}
\arguments{
\item{df}{the data frame}

\item{x_fac_vec}{the vector of factors to go on the x-axis.}

\item{y_fac_vec}{the vector of factors to go on the y-axis.}

\item{response_var}{the name of the response variable.}

\item{low_color}{the color for the lowest value.}

\item{mid_color}{the color for the midpoint value.}

\item{high_color}{the color for the highest value.}

\item{midpoint}{the numerical value of the middle point to color mid_color}

\item{digits}{the number of digits to round numbers in the plot to. Defaults to 2.}

\item{use_margins}{if TRUE, provide row and column margin aggregates. FALSE by
default.}

\item{tile_text_size}{the size of the text in tiles.
can be a number or a rel(1.0) for relative sizes.}
}
\value{
mizr_plot the plot object
}
\description{
Produce a Colored Table of all factors
}
