% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_tile_tf_plot}
\alias{mizr_tile_tf_plot}
\title{Produce a title table of all two-factor combinations. Used to provide
information on two-term interactions}
\usage{
mizr_tile_tf_plot(
  df,
  fac_vec,
  response_var,
  low_color = "#FF0000",
  mid_color = "#FFFFFF",
  high_color = "#0074FF",
  midpoint = 0,
  digits = 2,
  tile_text_size = rel(2),
  remove_NA_factors = FALSE
)
}
\arguments{
\item{df}{the data frame}

\item{fac_vec}{the vector of factors to do two-factor counts of}

\item{response_var}{the response variable column.}

\item{low_color}{the color for the lowest value.}

\item{mid_color}{the color for the midpoint value.}

\item{high_color}{the color for the highest value.}

\item{midpoint}{the numerical value of the middle point to color mid_color}

\item{digits}{the number of digits to round numbers in the plot to. Defaults to 2.}

\item{tile_text_size}{the size of the text in tiles.
can be a number or a rel(1.0) for relative sizes.}

\item{remove_NA_factors}{if TRUE, remove all attributes with a NA factor. FALSE by
default.}

\item{use_margins}{if TRUE, provide row and column margin aggregates. FALSE by
default.}
}
\value{
mizr_plot the plot object
}
\description{
Produce a title table of all two-factor combinations. Used to provide
information on two-term interactions
}
