% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_frames.R
\name{mizr_weighted_to_unweighted}
\alias{mizr_weighted_to_unweighted}
\title{Provides an aggregated data frame that weighs examples by category}
\usage{
mizr_weighted_to_unweighted(input_df, x_col_vec, response_var)
}
\description{
Takes an unbalanced (or with unbalanced replactions)
}
