% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_frames.R
\name{mizr_wide_to_long}
\alias{mizr_wide_to_long}
\title{Converts a wide data frame to a long data frame.}
\usage{
mizr_wide_to_long(input_df, id_col_vec)
}
\arguments{
\item{input_df}{the wide data frame to make long}

\item{id_col_vec}{the vector of the id columns}
}
\value{
output_df the long data frame
}
\description{
Uses the reshape2 package
}
