% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youden_plots.R
\name{mizr_youden_plot}
\alias{mizr_youden_plot}
\title{Produces a Youden Plot.
If the factors have multiple levels, the dataset groups them into two levels separating
the first level from all of the others in a macro-averaged format}
\usage{
mizr_youden_plot(
  df,
  effects_var_vec,
  response_var,
  plot_NA_values = FALSE,
  is_two_level = FALSE
)
}
\arguments{
\item{df}{the data frame with the data}

\item{effects_var_vec}{the list of colum names of the effect factors, specified in order}

\item{response_var}{the column name of the response variable}

\item{is_two_level}{a boolean if all of the factors are two-level only. Defaults
to FALSE}
}
\value{
mizr_plot a plot object
}
\description{
Produces a Youden Plot.
If the factors have multiple levels, the dataset groups them into two levels separating
the first level from all of the others in a macro-averaged format
}
