% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_plot.R
\name{one_term_level_avg_df}
\alias{one_term_level_avg_df}
\title{Compute main effect category averages}
\usage{
one_term_level_avg_df(
  df,
  effects_var_vec,
  response_var,
  plot_NA_values = FALSE,
  alpha = 0.05
)
}
\description{
Table format: ("Factor", "Level", "<response_var>")
where the "<response_var>" is the average (arithmetic mean) for that category
}
