% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_plots.R
\name{save_plot}
\alias{save_plot}
\title{Exports the plot as an image file}
\usage{
save_plot(plot_obj, filepath, format = "png", width, height, units = "in")
}
\arguments{
\item{filepath}{the file path to save the file to. Path is assumed to exist.}

\item{format}{the image format to save as. It is preferred that the extension suggested here
matches the extension of the filename in the filepath variable}

\item{width}{the width to save the image as}

\item{height}{the height to save the image as}

\item{units}{the units the width and height are represented in. Inches ("in") by default.}

\item{plot_object}{the plot object to save}
}
\value{
(nothing returned)
}
\description{
Exports the plot as an image file
}
