% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plots.R
\name{scatter_doe_produce_plot}
\alias{scatter_doe_produce_plot}
\title{Produces a scatter plot that colors and shapes by the treatment factor}
\usage{
scatter_doe_produce_plot(
  mizr_df,
  labels_vec,
  fac_vec,
  response_var,
  mean_val,
  sd_df = NULL,
  treatment_fac = NULL,
  doe_plot_style = "enhanced"
)
}
\arguments{
\item{mizr_df}{the data frame}

\item{labels_vec}{the vector names of the labels}

\item{fac_vec}{the factors to plot on the x column}

\item{response_var}{the response variable to plot on the y axis}

\item{mean_val}{the mean value}

\item{treatment_fac}{the treatment factor. NULL if none.}

\item{doe_plot_style}{Specifies the Plot Style. Use "plain" to mimic the DOE Scatter Plot
in dataplot, and "enhanced" to use an enhanced version, and
"boxplot" to place the boxplot in front of the points, and "sd" to replace the boxplot
To produce a mean and standard deviation rather than quantiles. Defaults to "enhanced".}
}
\value{
plot_obj the plot object
}
\description{
Produces a scatter plot that colors and shapes by the treatment factor
}
