% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plots.R
\name{scatter_effect_produce_plot}
\alias{scatter_effect_produce_plot}
\title{Produces a scatter plot that does a main effect of x_var by response_var}
\usage{
scatter_effect_produce_plot(
  effect_df,
  x_var,
  response_var,
  comp_func = mean,
  ...
)
}
\arguments{
\item{effect_df}{the data frame}

\item{x_var}{the factor to plot on the x column}

\item{response_var}{the response variable to plot on the y axis}

\item{comp_func}{mean by defalut. The computation function used to aggregate multiple values
into a single value}
}
\value{
plot_obj the plot object
}
\description{
Produces a scatter plot that does a main effect of x_var by response_var
}
