% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plots.R
\name{scatter_produce_plot}
\alias{scatter_produce_plot}
\title{Produces a basic scatter plot that colors and shapes by the treatment factor}
\usage{
scatter_produce_plot(df, x_var, response_var, shape_style = "point")
}
\arguments{
\item{df}{the data frame}

\item{x_var}{the factor to plot on the x column}

\item{response_var}{the response variable to plot on the y axis}

\item{shape_style}{a string indicating if the shape should be points ('point') or textual
symbols 'txt'. Defaults to point.}
}
\value{
plot_obj the plot object
}
\description{
Produces a basic scatter plot that colors and shapes by the treatment factor
}
