% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_plots.R
\name{sign_test_compute_df}
\alias{sign_test_compute_df}
\title{Computes a sign test rank df from the block plot}
\usage{
sign_test_compute_df(
  df,
  x_var_vec,
  treatment_var,
  response_var,
  higher_is_better = TRUE,
  use_margins = TRUE,
  alpha = 0.05,
  p_adj = "bonferonni",
  comp_func = mean,
  ...
)
}
\description{
Computes a sign test rank df from the block plot
}
