% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_plots.R
\name{sign_test_produce_plot}
\alias{sign_test_produce_plot}
\title{Produces a sign test rank plot}
\usage{
sign_test_produce_plot(
  sign_test_df,
  x_var_vec,
  treatment_var,
  response_var,
  use_margins = FALSE,
  include_equal_comparisons = TRUE,
  alpha = 0.05,
  p_adj = "bonferonni",
  low_color = "#FF0000",
  mid_color = "#FFFFFF",
  high_color = "#0074FF",
  tile_text_size = rel(2)
)
}
\description{
Produces a sign test rank plot
}
