% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_plots.R
\name{view_plot}
\alias{view_plot}
\title{Show plot}
\usage{
view_plot(plot_obj)
}
\arguments{
\item{plot_obj}{the plot object to plot}
}
\value{
plot_obj the same objected plotted so that it can be piped to a save_plot method
}
\description{
Show plot
}
