% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youden_plots.R
\name{youden_compute_df}
\alias{youden_compute_df}
\title{Produces a Youden Plot on datasets.
If the factors are multi-level, the data arbitrarily cuts the data into the
first field, and all the other levels (including NA)}
\usage{
youden_compute_df(
  df,
  effects_var_vec,
  response_var,
  plot_NA_values = FALSE,
  is_two_level = FALSE
)
}
\description{
Produces a Youden Plot on datasets.
If the factors are multi-level, the data arbitrarily cuts the data into the
first field, and all the other levels (including NA)
}
