% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youden_plots.R
\name{youden_produce_plot}
\alias{youden_produce_plot}
\title{Produces a main effects plot}
\usage{
youden_produce_plot(
  effects_all_df,
  mean_val,
  effects_var_vec,
  response_var,
  plot_NA_values = FALSE
)
}
\arguments{
\item{effects_var_vec}{the list of colum names of the effect factors, specified in order}

\item{response_var}{the column name of the response variable}

\item{df}{the data frame with the data}

\item{bar_ci}{If TRUE, plot bars that are the standard erro.
FALSE by default. Here sdn is computed similar to a standard error on the mean.}
}
\value{
plot_obj a plot object
}
\description{
Produces a main effects plot
}
