% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizr_plots.R
\name{mizr_block_plot}
\alias{mizr_block_plot}
\title{Produces a block plot}
\usage{
mizr_block_plot(
  df,
  x_var_vec,
  treatment_var,
  response_var,
  value_text_size = rel(1.6),
  comp_func = mean,
  ...
)
}
\arguments{
\item{df}{the data frame}

\item{x_var_vec}{the vector of columns to display on the x-axis.
Multiple factors can be combined to enumerate out the different values.}

\item{treatment_var}{the blocking factor or effect to plot separately as different symbols}

\item{response_var}{the response column to plot on the y axis;}

\item{value_text_size}{the size of the text within the plot. Often specified
as rel(Y) or a relative text size}

\item{comp_func}{mean by default. The computation function used to aggregate the multiple
values into a single value.}
}
\value{
mizr_plot a plot object with this plot to display
}
\description{
Produces a block plot
}
